move FrameRange to luxtest_utils
and add .iter_frames() method
